= SipUserAgent =  * x64 에서 컴파일할 때에 warning 이 나오지 않도록 수정함

+++ trunk/SipUserAgent/SipUserAgentCallBack.cpp
@@ -51,7 +51,7 @@ int CSipCallRtp::GetMediaCount( )
 	int iCount = 1;
 
 #ifdef USE_MEDIA_LIST
-	iCount = m_clsMediaList.size();
+	iCount = (int)m_clsMediaList.size();
 	if( iCount == 0 ) iCount = 1;
 #endif
